+++ game_patch/hud/hud_weapons.cpp
@@ -41,7 +41,7 @@ CallHook<void(int, int, int, rf::gr::Mode)> render_reticle_gr_bitmap_hook{
     },
     [](int bm_handle, int x, int y, rf::gr::Mode mode) {
         float base_scale = g_alpine_game_config.big_hud ? 2.0f : 1.0f;
-        float scale = base_scale * g_game_config.reticle_scale;
+        float scale = base_scale * g_alpine_game_config.reticle_scale;
         int clip_w = rf::gr::clip_width();
         int clip_h = rf::gr::clip_height();
 
@@ -89,12 +89,11 @@ ConsoleCommand2 reticle_scale_cmd{
     "ui_reticlescale",
     [](std::optional<float> scale_opt) {
         if (scale_opt) {
-            g_game_config.reticle_scale = scale_opt.value();
-            g_game_config.save();
+            g_alpine_game_config.set_reticle_scale(scale_opt.value());
         }
-        rf::console::print("Reticle scale {:.4f}", g_game_config.reticle_scale.value());
+        rf::console::print("Reticle scale {:.4f}", g_alpine_game_config.reticle_scale);
     },
-    "Sets/gets reticle scale",
+    "Set reticle scale",
 };
 
 bool hud_weapons_is_double_ammo()