+++ game_patch/object/obj_light.cpp
@@ -183,7 +183,7 @@ ConsoleCommand2 mesh_static_lighting_cmd{
 CallHook<void(rf::Entity&)> entity_update_muzzle_flash_light_hook{
     0x0041E814,
     [](rf::Entity& ep) {
-        if (g_game_config.try_disable_muzzle_flash && !server_side_restrict_disable_muzzle_flash) {
+        if (g_alpine_game_config.try_disable_muzzle_flash_lights && !server_side_restrict_disable_muzzle_flash) {
             return;
         }
         entity_update_muzzle_flash_light_hook.call_target(ep);
@@ -196,7 +196,7 @@ void evaluate_restrict_disable_muzzle_flash()
         rf::is_multi && !rf::is_server && get_df_server_info() && !get_df_server_info()->allow_no_mf;
 
     if (server_side_restrict_disable_muzzle_flash) {
-        if (g_game_config.try_disable_muzzle_flash) {
+        if (g_alpine_game_config.try_disable_muzzle_flash_lights) {
             rf::console::print("This server does not allow you to disable muzzle flash lights!");
         }
     }
@@ -205,13 +205,12 @@ void evaluate_restrict_disable_muzzle_flash()
 ConsoleCommand2 muzzle_flash_cmd{
     "r_muzzleflash",
     []() {
-        g_game_config.try_disable_muzzle_flash = !g_game_config.try_disable_muzzle_flash;
-        g_game_config.save();
+        g_alpine_game_config.try_disable_muzzle_flash_lights = !g_alpine_game_config.try_disable_muzzle_flash_lights;
 
         evaluate_restrict_disable_muzzle_flash();
 
         rf::console::print("Muzzle flash lights are {}",
-                           g_game_config.try_disable_muzzle_flash
+                           g_alpine_game_config.try_disable_muzzle_flash_lights
                                ? "disabled. In multiplayer, this will only apply if the server allows it."
                                : "enabled.");
     },