+++ launcher/OptionsMiscDlg.cpp
@@ -18,7 +18,6 @@ BOOL OptionsMiscDlg::OnInitDialog()
     CheckDlgButton(IDC_PLAYER_JOIN_BEEP_CHECK, m_conf.player_join_beep ? BST_CHECKED : BST_UNCHECKED);
     CheckDlgButton(IDC_AUTOSAVE_CHECK, m_conf.autosave ? BST_CHECKED : BST_UNCHECKED);
     CheckDlgButton(IDC_ALPINE_BRAND_CHECK, m_conf.af_branding ? BST_CHECKED : BST_UNCHECKED);
-    CheckDlgButton(IDC_MUZZLEFLASH, m_conf.try_disable_muzzle_flash ? BST_CHECKED : BST_UNCHECKED);
     CheckDlgButton(IDC_DAMAGEFLASH, m_conf.damage_screen_flash ? BST_CHECKED : BST_UNCHECKED);
     CheckDlgButton(IDC_POW2TEX, m_conf.pow2tex ? BST_CHECKED : BST_UNCHECKED);
 
@@ -34,7 +33,6 @@ void OptionsMiscDlg::InitToolTip()
     m_tool_tip.AddTool(GetDlgItem(IDC_PLAYER_JOIN_BEEP_CHECK), "Play a beep when a player joins the server you are in while your game doesn't have focus");
     m_tool_tip.AddTool(GetDlgItem(IDC_AUTOSAVE_CHECK), "Automatically save the game after a level transition");
     m_tool_tip.AddTool(GetDlgItem(IDC_ALPINE_BRAND_CHECK), "Display Alpine Faction branding where available");
-    m_tool_tip.AddTool(GetDlgItem(IDC_MUZZLEFLASH), "Disable blinking dynamic lights when players are shooting (in multiplayer, only applies if the server allows it)");
     m_tool_tip.AddTool(GetDlgItem(IDC_DAMAGEFLASH), "Show a bright red screen flash when you suffer damage");
     m_tool_tip.AddTool(GetDlgItem(IDC_POW2TEX), "Enforce for textures with nonstandard resolutions, fixes textures in older levels but may cause issues with textures in newer ones");
 }
@@ -47,7 +45,6 @@ void OptionsMiscDlg::OnSave()
     m_conf.player_join_beep = (IsDlgButtonChecked(IDC_PLAYER_JOIN_BEEP_CHECK) == BST_CHECKED);
     m_conf.autosave = (IsDlgButtonChecked(IDC_AUTOSAVE_CHECK) == BST_CHECKED);
     m_conf.af_branding = (IsDlgButtonChecked(IDC_ALPINE_BRAND_CHECK) == BST_CHECKED);
-    m_conf.try_disable_muzzle_flash = (IsDlgButtonChecked(IDC_MUZZLEFLASH) == BST_CHECKED);
     m_conf.damage_screen_flash = (IsDlgButtonChecked(IDC_DAMAGEFLASH) == BST_CHECKED);
     m_conf.pow2tex = (IsDlgButtonChecked(IDC_POW2TEX) == BST_CHECKED);
 }