+++ launcher/resource.h
@@ -78,7 +78,7 @@
 //#define IDC_FULLBRIGHT                  1060
 //#define IDC_LIGHTMAPS                   1061
 #define IDC_NEAREST                     1062
-#define IDC_MUZZLEFLASH                 1063
+//#define IDC_MUZZLEFLASH                 1063
 //#define IDC_SCREENSHAKE                 1064
 #define IDC_DAMAGEFLASH                 1065
 #define IDC_POW2TEX                     1066