+++ docs/CHANGELOG.md
@@ -100,6 +100,7 @@ Version 1.1.0 (Tamarack): Not yet released
   - Scanner sensitivity modifier, scope sensitivity modifier, static scope sensitivity
   - World HUD objective icons, overdraw, big text, damage numbers, spectate labels, team labels
   - Hit sounds and audio taunts
+  - Muzzle flash lights, unlimited semi auto, gaussian spread, ricochet
 - Add `Alpine Faction` options panel to in-game options menu with settings:
     - Big HUD, linear pitch, swap controls for AR, SG, GN
     - Weapon shake, fullbright characters, lightmaps only